NBPB Framework Source Code Draft/requirements.txt
# requirements.txt
# Core dependencies
torch>=1.11.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
pandas>=1.3.0
PyYAML>=6.0
# Optional but recommended
matplotlib>=3.4.0
seaborn>=0.11.0
 